% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hw04pSantinule-package.R
\docType{package}
\name{hw04pSantinule-package}
\alias{hw04pSantinule}
\alias{hw04pSantinule-package}
\title{hw04pSantinule: Has awesome functions}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Santiago Nule \email{sn1964a@student.american.edu} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
