% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/santifunc1.R
\name{santifunc1}
\alias{santifunc1}
\title{Returns the nth term from the recursive sequence xn=xn−1+(xn−3−xn−2)/n.}
\usage{
santifunc1(x, n)
}
\arguments{
\item{x}{a vector of length 3}

\item{n}{the number of terms for the recursive sequence}
}
\value{
returns the nth term from the recursive sequence
}
\description{
Returns the nth term from the recursive sequence xn=xn−1+(xn−3−xn−2)/n.
}
\examples{
santifunc1(x = c(2, 4, 3), n = 0)
santifunc1(x = c(2, 4, 3), n = 6)
}
