% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/santifunc2.R
\name{santifunc2}
\alias{santifunc2}
\title{Returns a line graph for multiple values that go into santifunc1}
\usage{
santifunc2(df)
}
\arguments{
\item{df}{A data frame that contains four columns. The first three columns are the values of the three numerics to be input to function 1 and the fourth column is the positive integer n for the sequence to be generated.}
}
\value{
Returns a line graph for multiple values that go into santifunc1
}
\description{
Returns a line graph for multiple values that go into santifunc1
}
\examples{

}
